fix: #WB2-1514, fix feedback PR
